Refactor for improved efficiency and simpler code Improved efficiency of setting saved states O(n^2) to O(n)
