+++ src/Model/AutoSegmentation/AutoSegmentViewState.py
@@ -1,0 +1,3 @@
+class AutoSegmentViewState:
+    def __init__(self):
+        self.segmentation_list = []
